refactor: split amber data serializer
